Huge client class disassembled into small ones, removed unnecessary types, removed warnings
