+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.redis</groupId>
     <artifactId>redis-scala-netty</artifactId>
-    <version>1.1.7</version>
+    <version>1.2.0</version>
     <packaging>jar</packaging>
 
     <properties>
@@ -21,12 +21,6 @@
         </dependency -->
 
         <dependency>
-            <groupId>org.scala-lang</groupId>
-            <artifactId>scala-actors</artifactId>
-            <version>${scala.version}</version>
-        </dependency>
-            
-        <dependency>
             <groupId>io.netty</groupId>
             <artifactId>netty</artifactId>
             <version>3.9.3.Final</version>
@@ -36,6 +30,13 @@
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
             <version>4.11</version>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.scala-lang</groupId>
+            <artifactId>scala-library</artifactId>
+            <version>${scala.version}</version>
             <scope>test</scope>
         </dependency>
